brain/pyproject.toml
[project]
name = "tekno-phantom-brain"
version = "0.1.0"
description = "Standalone LLM brain service for Tekno Phantom Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "openai>=1.50.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
